House, powerup, difficulty scale
Added a house asset on one end of the road and complete the layout with the road and house on one end of the street. also slanted the 2D camera view so you can see the house. Added a powerup that spawns every 25 seconds, it's an ice ball that destroys all cars on the road and slows down the ones spawning for 5 seconds. Also included an update for difficulty increasing overtime. It still needs to be adjusted and balanced.